modified:   src/aces/aces_library/special_directory/direct_integral/binpack.f  	modified:   src/aces/aces_library/special_directory/direct_integral/ecp_renorm.F  	modified:   src/aces/aces_sial/scf/scf_aguess.sial  	modified:   src/aces/aces_sial/sial_config  	modified:   src/main/det_segsizes.F  	modified:   src/sia/sip/aces_instructions/compute_atomscf.f  	modified:   src/sia/sip/aces_instructions/diag.F  	modified:   src/sia/sip_shared/mem_alloc.c  	modified:   src/sia/worker/determine_stack_blocksizes.f  	src/aces/aces_sial/eomccsdpt/  	src/sia/sip/aces_instructions/diag.F.nolindep

+++ src/sia/sip_shared/mem_alloc.c
@@ -70,7 +70,7 @@ void F77_NAME(mem_alloc,MEM_ALLOC)
 	assert (total_bytes != 0);
 	*ierr = 0;
     f_int esize = *element_size;
-	long long nbytes = (*nwords) * esize;
+	long long nbytes = (long long) *nwords * esize;
 	long long fiddle = ALIGN-1;
 	long long nxt_int = (long long)nxt_ptr;
 //        printf("mem_alloc called with nwords %d and esize %d, previous nxt_int is %lld.\n",*nwords,esize,nxt_int);
